Workaround CocoaPods not validating us
For some reason CocoaPods wasn’t validating us if we disabled this warning.

We either disable this warning or have to put every method into its own category, and while that’s OK for now, I’m not keen on having to remember an extra step for every new method I add to the library. Especially since the warning is not an issue for our usage. This fixes that.